feat: add workflow TAT actions and queries for ER and triage cleaning stages
