+++ .env.example
@@ -142,44 +142,7 @@ BACKUP_RETENTION_DAYS=30
 # Generate encryption key: node scripts/encrypt-secret.js
 # See CONFIGURATION.md for encryption setup details
 
-# ============================================================================
-# ENVIRONMENT-SPECIFIC CONFIGURATION EXAMPLES
-# ============================================================================
-
-# ----------------------------------------------------------------------------
-# LOCAL DEVELOPMENT (.env.local)
-# ----------------------------------------------------------------------------
-# DATABASE_URL=postgresql://postgres:yourpassword@localhost:5432/ewtcs_dev
-# SESSION_SECRET=dev-secret-at-least-32-characters-long-please
-# NEXT_PUBLIC_APP_URL=http://localhost:3000
-# PORT=3000
-# NODE_ENV=development
-# RED_ALERT_THRESHOLD_MS=10800000
-
-# ----------------------------------------------------------------------------
-# STAGING ENVIRONMENT (.env.staging)
-# ----------------------------------------------------------------------------
-# DATABASE_URL=postgresql://ewtcs_user:strong_password@staging-db.example.com:5432/ewtcs_staging
-# SESSION_SECRET=staging-secret-key-at-least-32-characters-long
-# NEXT_PUBLIC_APP_URL=https://staging.yourdomain.com
-# PORT=3000
-# NODE_ENV=staging
-# RED_ALERT_THRESHOLD_MS=10800000
-
-# ----------------------------------------------------------------------------
-# PRODUCTION ENVIRONMENT (.env.production)
-# ----------------------------------------------------------------------------
-# Use encrypted secrets in production for security
-# DATABASE_URL_ENCRYPTED=ivhex:your-encrypted-database-url-here
-# ENCRYPTION_KEY=your-64-character-hex-encryption-key-here
-# SESSION_SECRET=production-secret-key-minimum-32-characters-long
-# NEXT_PUBLIC_APP_URL=https://yourdomain.com
-# PORT=3000
-# NODE_ENV=production
-# RED_ALERT_THRESHOLD_MS=10800000
-
-# ============================================================================
-# ============================================================================
+# Environment-specific examples live in CONFIGURATION.md.
 # ENCRYPTION CONFIGURATION (US-18: Data Encryption & Sensitive Data Protection)
 # ============================================================================
 # Master encryption key for protecting sensitive data at rest
