+++ src/features/ai-summary/lib/daily-core-aggregation-queries.ts
@@ -1,0 +1,79 @@
+import { query } from '@/shared/lib/db'
+import type {
+  RawAvgStageTime,
+  RawDelayCount,
+  RawMostDelayedStage,
+  RawPatientStats,
+} from './aggregation-helpers'
+
+export async function getPatientStats(
+  dayStart: Date,
+  dayEnd: Date
+): Promise<{ totalPatients: number; totalBedsUsed: number; totalStageUpdates: number }> {
+  const result = await query<RawPatientStats>(
+    `
+    SELECT
+      COUNT(DISTINCT pa.id) AS "totalPatients",
+      COUNT(DISTINCT bsl.bed_id) AS "totalBedsUsed",
+      COUNT(bsl.id) AS "totalStageUpdates"
+    FROM bed_stage_logs bsl
+    LEFT JOIN patient_admissions pa
+      ON pa.bed_id = bsl.bed_id
+     AND pa.admitted_at >= $1
+     AND pa.admitted_at <= $2
+    WHERE bsl.transition_time >= $1
+      AND bsl.transition_time <= $2
+  `,
+    [dayStart, dayEnd]
+  )
+  const row = result.rows[0]
+  return {
+    totalPatients: parseInt(row?.totalPatients ?? '0', 10),
+    totalBedsUsed: parseInt(row?.totalBedsUsed ?? '0', 10),
+    totalStageUpdates: parseInt(row?.totalStageUpdates ?? '0', 10),
+  }
+}
+
+export async function getAvgStageTime(dayStart: Date, dayEnd: Date): Promise<number> {
+  const result = await query<RawAvgStageTime>(
+    `
+    SELECT AVG(bsl.duration_in_previous_stage_ms) AS "avgStageTimeMs"
+    FROM bed_stage_logs bsl
+    WHERE bsl.transition_time >= $1 AND bsl.transition_time <= $2
+      AND bsl.duration_in_previous_stage_ms IS NOT NULL
+  `,
+    [dayStart, dayEnd]
+  )
+  const raw = result.rows[0]?.avgStageTimeMs
+  return raw !== null && raw !== undefined ? parseFloat(raw) : 0
+}
+
+export async function getDelayCount(dayStart: Date, dayEnd: Date): Promise<number> {
+  const result = await query<RawDelayCount>(
+    `
+    SELECT COUNT(DISTINCT ddr.bed_id) AS "delayCount"
+    FROM disposition_delay_reasons ddr
+    WHERE ddr.recorded_at >= $1 AND ddr.recorded_at <= $2
+  `,
+    [dayStart, dayEnd]
+  )
+  return parseInt(result.rows[0]?.delayCount ?? '0', 10)
+}
+
+export async function getMostDelayedStage(
+  dayStart: Date,
+  dayEnd: Date
+): Promise<string | undefined> {
+  const result = await query<RawMostDelayedStage>(
+    `
+    SELECT s.name AS "stageName"
+    FROM disposition_delay_reasons ddr
+    JOIN bed_stage_logs bsl ON bsl.id = ddr.bed_stage_log_id
+    JOIN stages s ON s.id = bsl.to_stage_id
+    WHERE ddr.recorded_at >= $1 AND ddr.recorded_at <= $2
+    GROUP BY s.name ORDER BY COUNT(*) DESC LIMIT 1
+  `,
+    [dayStart, dayEnd]
+  )
+  return result.rows[0]?.stageName ?? undefined
+}
